+++ container-engine-lib/docs/changelog.md
@@ -17,7 +17,7 @@
     * Users should use `DestroyUserServices`
 * `CreateUserService` has been renamed to `StartUserService`
     * TODO
-* TODO scoping down the service requests to be per-enclave
+* TODO scoping down all Service functions to be per-enclave
 * TODO remove wait-for-availability because it doesn't need to be there
 
 # 0.21.1